Rollout table: Display tooltip for status icons Disable update option for rollout in creatinga nd starting state

+++ hawkbit-repository/src/main/java/org/eclipse/hawkbit/repository/RolloutManagement.java
@@ -346,7 +346,7 @@ public class RolloutManagement {
         int pageIndex = 0;
         int groupIndex = 0;
         final Long totalCount = savedRollout.getTotalTargets();
-        final int groupSize = (int) Math.ceil(totalCount / amountGroup);
+        final int groupSize = (int) Math.ceil((double) totalCount / (double) amountGroup);
         RolloutGroup lastSavedGroup = null;
         while (pageIndex < totalCount) {
             groupIndex++;
